+++ src/mruby_integration/image.cpp
@@ -236,6 +236,15 @@ mrb_value mrb_image_colour_invert(mrb_state *mrb, mrb_value) {
   return mrb_nil_value();
 }
 
+mrb_value mrb_image_colour_grayscale(mrb_state *mrb, mrb_value) {
+  Image *image;
+  mrb_get_args(mrb, "d", &image, &Image_type);
+
+  ImageColorGrayscale(image);
+
+  return mrb_nil_value();
+}
+
 mrb_value mrb_image_resize_nearest_neighbour(mrb_state *mrb, mrb_value) {
   mrb_value image_obj;
   mrb_int width, height;
@@ -292,6 +301,7 @@ void append_images(mrb_state *mrb) {
 
   mrb_define_method(mrb, mrb->kernel_module, "image_colour_tint!", mrb_image_colour_tint, MRB_ARGS_REQ(2));
   mrb_define_method(mrb, mrb->kernel_module, "image_colour_invert!", mrb_image_colour_invert, MRB_ARGS_REQ(1));
+  mrb_define_method(mrb, mrb->kernel_module, "image_colour_grayscale!", mrb_image_colour_grayscale, MRB_ARGS_REQ(1));
 
   mrb_define_method(mrb, mrb->kernel_module, "generate_image_colour", mrb_generate_image_colour, MRB_ARGS_REQ(3));